Added in the docker compose files: 1. MQTT broker for communication with Temi 2. Python webserver to make video files in new videos folder accessible Also adjusted some configurations for the custom MQTT broker

+++ node-red/npmInstall.sh
@@ -4,9 +4,13 @@
 packages=(./nodes/pepper/ ./nodes/temi/ ./nodes/sawyer/ ./nodes/base/)
 
 for i in "${packages[@]}"; do
-   :
-   # npm install $i --no-audit --progress=false --no-fund
-   npm install $i --no-audit --progress=false
+   if [ -d "$i" ] && [ -f "$i/package.json" ]; then
+
+      # npm install $i --no-audit --progress=false --no-fund
+      npm install $i --no-audit --progress=false
+   else
+      echo "Skipping $i (directory package.json not found)"
+   fi
 done
 
 node-red start -v --userDir /data $FLOWS
